+++ backend/convertform/src/main/java/com/example/convertform/entity/User.java
@@ -2,8 +2,6 @@ package com.example.convertform.entity;
 
 import lombok.Data;
 
-import java.util.List;
-
 @Data
 public class User {
     private Integer id;
