+++ backend/convertform/src/main/java/com/example/convertform/jwt/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.example.convertform.jwt;
 
-import com.example.convertform.service.impl.repository.UserService;
+import com.example.convertform.service.impl.storage.UserService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
