+++ backend/convertform/src/main/java/com/example/convertform/service/impl/storage/UserService.java
@@ -1,6 +1,5 @@
-package com.example.convertform.service.impl.repository;
+package com.example.convertform.service.impl.storage;
 
-import com.example.convertform.entity.CustomUserDetails;
 import com.example.convertform.entity.User;
 import com.example.convertform.sqlmapper.UserMapper;
 import lombok.RequiredArgsConstructor;
